完成运行bot代码的微服务，并编写简易AI

+++ backend/main-server/src/main/java/com/kob/backend/mainserver/pojo/Record.java
@@ -21,13 +21,13 @@ import java.util.Date;
 public class Record {
 
     @TableId(type = IdType.AUTO)
-    private Long id;
+    private Integer id;
 
-    private Long aId;
+    private Integer aId;
 
     private Integer aX, aY;
 
-    private Long bId;
+    private Integer bId;
 
     private Integer bX, bY;
